+++ Demo/Demo/ContentView.swift
@@ -3,7 +3,7 @@
 //  ScrollKit
 //
 //  Created by Daniel Saidi on 2023-02-04.
-//  Copyright © 2022 Daniel Saidi. All rights reserved.
+//  Copyright © 2023 Daniel Saidi. All rights reserved.
 //
 
 import SwiftUI
@@ -11,34 +11,24 @@ import ScrollKit
 
 struct ContentView: View {
 
-    @State
-    private var isStatusBarHidden = false
+    @StateObject
+    private var statusBarVisibile = StatusBarVisibileState()
 
     var body: some View {
         NavigationStack {
-            VStack {
-                List {
-                    title
-                    linkSection
-                }
+            List {
+                linkSection
             }
             .tint(.blue)
-            .toolbar(.hidden)
+            .navigationTitle("Demo")
         }
         .tint(.white)
         .toolbarRole(.navigationStack)
-        .statusBar(hidden: isStatusBarHidden)
+        .statusBarVisibile(statusBarVisibile)
     }
 }
 
 private extension ContentView {
-
-    var title: some View {
-        Text("ScrollKit Demo")
-            .font(.largeTitle)
-            .listRowInsets(.none)
-            .listRowBackground(Color.clear)
-    }
 
     var linkSection: some View {
         Section(header: Text("Stretchable headers")) {
@@ -56,7 +46,7 @@ private extension ContentView {
 
     var colorLink: some View {
         link("paintbrush.pointed.fill", "Short color background") {
-            DemoScreen(isStatusBarHidden: $isStatusBarHidden, headerHeight: 100) {
+            DemoScreen(headerHeight: 100) {
                 Color.blue
             }
         }
@@ -64,7 +54,7 @@ private extension ContentView {
 
     var gradientLink: some View {
         link("paintbrush.fill", "Gradient background") {
-            DemoScreen(isStatusBarHidden: $isStatusBarHidden, headerHeight: 250) {
+            DemoScreen(headerHeight: 250) {
                 ScrollViewHeaderGradient(.yellow, .blue)
             }
         }
@@ -72,7 +62,7 @@ private extension ContentView {
 
     var imageLink: some View {
         link("photo.fill", "Image background") {
-            DemoScreen(isStatusBarHidden: $isStatusBarHidden, headerHeight: 250) {
+            DemoScreen(headerHeight: 250) {
                 ZStack {
                     ScrollViewHeaderImage(Image("header"))
                     ScrollViewHeaderGradient(.black.opacity(0.2), .black.opacity(0.5))